refactor(deepspeed): Refine traning code
