+++ wenet/utils/executor.py
@@ -18,7 +18,9 @@ from contextlib import nullcontext
 # if your python version < 3.7 use the below one
 # from contextlib import suppress as nullcontext
 import torch
-from torch.nn.utils import clip_grad_norm_
+
+from wenet.utils.train_utils import (batch_forward, batch_backward,
+                                     update_parameter_and_lr, log_per_step)
 
 
 class Executor:
@@ -31,25 +33,9 @@ class Executor:
         ''' Train one epoch
         '''
         model.train()
-        clip = args.get('grad_clip', 50.0)
-        log_interval = args.get('log_interval', 10)
-        rank = args.get('rank', 0)
-        epoch = args.get('epoch', 0)
-        accum_grad = args.get('accum_grad', 1)
-        is_distributed = args.get('is_distributed', True)
-        is_deepspeed = args.get('is_deepspeed', False)
-        use_amp = args.get('use_amp', False)
-        ds_dtype = args.get('ds_dtype', "fp32")
-        if ds_dtype == "fp16":
-            ds_dtype = torch.float16
-        elif ds_dtype == "bf16":
-            ds_dtype = torch.bfloat16
-        else:
-            ds_dtype = None
         logging.info('using accumulate grad, new batch size is {} times'
-                     ' larger than before'.format(accum_grad))
-        if use_amp:
-            assert scaler is not None
+                     ' larger than before'.format(args.get('accum_grad', 1)))
+
         # A context manager to be used in conjunction with an instance of
         # torch.nn.parallel.DistributedDataParallel to be able to train
         # with uneven inputs across participating processes.
@@ -57,148 +43,75 @@ class Executor:
             model_context = model.join
         else:
             model_context = nullcontext
-        num_seen_utts = 0
+
         with model_context():
             for batch_idx, batch in enumerate(data_loader):
                 key, feats, target, feats_lengths, target_lengths = batch
-                feats = feats.to(device)
-                target = target.to(device)
-                feats_lengths = feats_lengths.to(device)
-                target_lengths = target_lengths.to(device)
-                num_utts = target_lengths.size(0)
-                if num_utts == 0:
+
+                batch_dict = {}
+                batch_dict["feats"] = feats.to(device)
+                batch_dict["target"] = target.to(device)
+                batch_dict["feats_lengths"] = feats_lengths.to(device)
+                batch_dict["target_lengths"] = target_lengths.to(device)
+
+                if target_lengths.size(0) == 0:
                     continue
+
                 context = None
                 # Disable gradient synchronizations across DDP processes.
                 # Within this context, gradients will be accumulated on module
                 # variables, which will later be synchronized.
-                if is_distributed and batch_idx % accum_grad != 0:
+                if args.get("train_engine", "torch_ddp") == "torch_ddp" and \
+                        batch_idx % accum_grad != 0:
                     context = model.no_sync
                 # Used for single gpu training and DDP gradient synchronization
                 # processes.
                 else:
                     context = nullcontext
-                with context():
-                    if is_deepspeed:  # deepspeed
-                        with torch.cuda.amp.autocast(
-                            enabled=ds_dtype is not None,
-                            dtype=ds_dtype, cache_enabled=False
-                        ):
-                            loss_dict = model(feats, feats_lengths, target,
-                                              target_lengths)
-                        loss = loss_dict['loss']
-                        # NOTE(xcsong): Zeroing the gradients is handled automatically by DeepSpeed after the weights # noqa
-                        #   have been updated using a mini-batch. DeepSpeed also performs gradient averaging automatically # noqa
-                        #   at the gradient accumulation boundaries and addresses clip_grad_norm internally. In other words # noqa
-                        #   `model.backward(loss)` is equivalent to `loss.backward() + clip_grad_norm_() + optimizer.zero_grad() + accum_grad` # noqa
-                        #   ref: https://www.deepspeed.ai/tutorials/megatron/#using-the-training-api  # noqa
-                        model.backward(loss)
-                    else:             # pytorch native ddp
-                        # autocast context
-                        # The more details about amp can be found in
-                        # https://pytorch.org/docs/stable/notes/amp_examples.html
-                        with torch.cuda.amp.autocast(scaler is not None):
-                            loss_dict = model(feats, feats_lengths, target,
-                                              target_lengths)
-                            loss = loss_dict['loss'] / accum_grad
-                        if use_amp:
-                            scaler.scale(loss).backward()
-                        else:
-                            loss.backward()
 
-                num_seen_utts += num_utts
-                if is_deepspeed:
-                    if rank == 0 and writer is not None \
-                            and model.is_gradient_accumulation_boundary():
-                        writer.add_scalar('train_loss', loss.item(), self.step)
-                    # NOTE(xcsong): The step() function in DeepSpeed engine updates the model parameters as well as the learning rate. There is # noqa
-                    #   no need to manually perform scheduler.step(). In other words: `ds_model.step() = optimizer.step() + scheduler.step()` # noqa
-                    #   ref: https://www.deepspeed.ai/tutorials/megatron/#using-the-training-api  # noqa
-                    model.step()
-                    self.step += 1
-                elif not is_deepspeed and batch_idx % accum_grad == 0:
-                    if rank == 0 and writer is not None:
-                        writer.add_scalar('train_loss', loss, self.step)
-                    # Use mixed precision training
-                    if use_amp:
-                        scaler.unscale_(optimizer)
-                        grad_norm = clip_grad_norm_(model.parameters(), clip)
-                        # Must invoke scaler.update() if unscale_() is used in
-                        # the iteration to avoid the following error:
-                        #   RuntimeError: unscale_() has already been called
-                        #   on this optimizer since the last update().
-                        # We don't check grad here since that if the gradient
-                        # has inf/nan values, scaler.step will skip
-                        # optimizer.step().
-                        scaler.step(optimizer)
-                        scaler.update()
-                    else:
-                        grad_norm = clip_grad_norm_(model.parameters(), clip)
-                        if torch.isfinite(grad_norm):
-                            optimizer.step()
-                    optimizer.zero_grad()
-                    scheduler.step()
-                    self.step += 1
-                if batch_idx % log_interval == 0:
-                    lr = optimizer.param_groups[0]['lr']
-                    log_str = 'TRAIN Batch {}/{} loss {:.6f} '.format(
-                        epoch, batch_idx,
-                        loss.item() * accum_grad)
-                    for name, value in loss_dict.items():
-                        if name != 'loss' and value is not None:
-                            log_str += '{} {:.6f} '.format(name, value.item())
-                    log_str += 'lr {:.8f} rank {}'.format(lr, rank)
-                    logging.debug(log_str)
+                with context():
+                    loss_dict = batch_forward(args, model, batch_dict, scaler)
+                    batch_backward(args, model, loss_dict, scaler)
+
+                info_dict = {"lr": lr, "batch_idx": batch_idx, "step": self.step}
+
+                lr = update_parameter_and_lr(
+                    args, model, optimizer, scheduler,
+                    scaler, info_dict
+                )
+
+                log_per_step(args, loss_dict, info_dict, writer, tag="TRAIN")
+
+                self.step += 1
 
     def cv(self, model, data_loader, device, args):
         ''' Cross validation on
         '''
         model.eval()
-        rank = args.get('rank', 0)
-        epoch = args.get('epoch', 0)
-        log_interval = args.get('log_interval', 10)
-        is_deepspeed = args.get('is_deepspeed', False)
-        ds_dtype = args.get('ds_dtype', "fp32")
-        if ds_dtype == "fp16":
-            ds_dtype = torch.float16
-        elif ds_dtype == "bf16":
-            ds_dtype = torch.bfloat16
-        else:  # fp32
-            ds_dtype = None
         # in order to avoid division by 0
         num_seen_utts = 1
         total_loss = 0.0
         with torch.no_grad():
             for batch_idx, batch in enumerate(data_loader):
                 key, feats, target, feats_lengths, target_lengths = batch
-                feats = feats.to(device)
-                target = target.to(device)
-                feats_lengths = feats_lengths.to(device)
-                target_lengths = target_lengths.to(device)
+
+                batch_dict = {}
+                batch_dict["feats"] = feats.to(device)
+                batch_dict["target"] = target.to(device)
+                batch_dict["feats_lengths"] = feats_lengths.to(device)
+                batch_dict["target_lengths"] = target_lengths.to(device)
+
                 num_utts = target_lengths.size(0)
                 if num_utts == 0:
                     continue
-                if is_deepspeed:
-                    with torch.cuda.amp.autocast(
-                        enabled=ds_dtype is not None,
-                        dtype=ds_dtype, cache_enabled=False
-                    ):
-                        loss_dict = model(feats, feats_lengths,
-                                          target, target_lengths)
-                else:
-                    loss_dict = model(feats, feats_lengths, target, target_lengths)
+
+                loss_dict = batch_forward(args, model, batch_dict, None)
                 loss = loss_dict['loss']
+
                 if torch.isfinite(loss):
                     num_seen_utts += num_utts
                     total_loss += loss.item() * num_utts
-                if batch_idx % log_interval == 0:
-                    log_str = 'CV Batch {}/{} loss {:.6f} '.format(
-                        epoch, batch_idx, loss.item())
-                    for name, value in loss_dict.items():
-                        if name != 'loss' and value is not None:
-                            log_str += '{} {:.6f} '.format(name, value.item())
-                    log_str += 'history loss {:.6f}'.format(total_loss /
-                                                            num_seen_utts)
-                    log_str += ' rank {}'.format(rank)
-                    logging.debug(log_str)
+
+                info_dict = {"batch_idx": batch_idx, "history_loss": total_loss / num_seen_utts}  # noqa
+                log_per_step(args, loss_dict, info_dict, writer, tag="CV")
         return total_loss, num_seen_utts
